+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 Voucherify Java SDK - Changelog
 ===============================
 
-- **2016-12-02** - `3.5.1` - Added gift balance. Enhanced utils to support gift vouchers.
+- **2016-12-02** - `3.6.1` - Added gift balance. Enhanced utils to support gift vouchers.
+- **2016-09-06** - `3.6.0` - Added a method to publish voucher for given code or campaign name.
 - **2016-09-06** - `3.5.0` - Added order items.
 - **2016-07-19** - `3.4.0` - Voucher code config.
 - **2016-07-18** - `3.3.0` - Added a method to update voucher.